fastapi-booking/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt>=4.1.0
pydantic-settings==2.6.1
python-multipart==0.0.17
email-validator==2.2.0
